+++ weather@mockturtl/files/weather@mockturtl/3.8/main.js
@@ -11,7 +11,6 @@ const darkSky_1 = require("providers/darkSky");
 const openWeatherMap_1 = require("providers/openWeatherMap");
 const us_weather_1 = require("providers/us_weather");
 const weatherbit_1 = require("providers/weatherbit");
-const yahoo_1 = require("providers/yahoo");
 const met_norway_1 = require("providers/met_norway");
 const httpLib_1 = require("lib/httpLib");
 const logger_1 = require("lib/logger");
@@ -312,10 +311,6 @@ class WeatherApplet extends TextIconApplet {
             case "Weatherbit":
                 if (currentName != "Weatherbit" || force)
                     this.provider = new weatherbit_1.Weatherbit(this);
-                break;
-            case "Yahoo":
-                if (currentName != "Yahoo" || force)
-                    this.provider = new yahoo_1.Yahoo(this);
                 break;
             case "ClimacellV4":
                 if (currentName != "ClimacellV4" || force)